Moved server code to explicit binary file
This change also adds an explicit binary file for
client code, but this code currently does not
compile. The issue is that the two tasks spawned
to send commands using the client both need to
access the client, which does not implement Copy.

Additionally, the call to `set` requires exclusive
access to client as it takes `&mut self`, meaning
the other task will not be able to borrow `client`
for as long as that acess is taken up.

As access to the client would have to persist
across calls to `await`, a mutex is not a good
solution to this problem. Instead, message passing
can be used to send commands to the client without
needing direct access to it.